fix: ViewModel Optional 구조로 변경
- 파일 분리
- NavigationableViewModel 개체 제거
