+++ Sources/ControllerableViewModel/StateView.swift
@@ -1,0 +1,14 @@
+//
+//  StateView.swift
+//
+//
+//  Created by 김용우 on 2023/10/17.
+//
+
+import SwiftUI
+
+open class StateView {
+    public weak var viewController: UIViewController?
+    
+    public init() {}
+}
