+++ Tests/ControllerableViewModelTests/ControllerableViewModelTests.swift
@@ -6,6 +6,6 @@ final class ControllerableViewModelTests: XCTestCase {
         // This is an example of a functional test case.
         // Use XCTAssert and related functions to verify your tests produce the correct
         // results.
-        XCTAssertEqual(ControllerableViewModel().text, "Hello, World!")
+        XCTAssertEqual("Hello, World!", "Hello, World!")
     }
 }
